requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
python-dotenv==1.0.1
yarl==1.9.4
